fix(windows): keep event props out of ntk's creation attributes
A `<popup>` whose props change before it is realized went down the
pre-realize branch of `WindowNode.applyProps`, which refreshes the
creation attributes rather than talking to a window that does not exist
yet. It spread the raw props in — including the event handlers.

ntk's Window constructor scans its creation args for `onKeyDown`,
`onFocus` & co. (events_map.toSnake) and registers each as a raw
listener. So the handler ran twice: once from the EventManager with the
synthetic event, and once from ntk with the native X event, which has no
`preventDefault`. That is the crash a Dialog's Escape key hit —
`ev.preventDefault is not a function`. It also pinned the first render's
closure for the window's lifetime.

`createInstance` already had the right filter, so this moves
`windowAttributes()` into nodes.js and runs the pre-realize update
through it. `transientFor` was deleted by hand there and is dropped by
the shared filter now.

A plain `<window>` realizes when it is appended, so it never took that
path; a popup realizes in commitMount, one update later.
