+++ src/Reconciler.js
@@ -29,9 +29,8 @@ import {
   TextAreaNode,
   flushWindowRestacks,
   flushTokenChecks,
-  WINDOW_HINT_PROPS,
+  windowAttributes,
 } from './nodes.js';
-import { flattenStyle } from './styles.js';
 import { hasDropProps } from './dnd.js';
 import { AppProvider } from './appcontext.js';
 import { defaultRootHandlers, setErrorHandler } from './errors.js';
@@ -87,41 +86,6 @@ export const HOST_TYPES = [
   'tex',
   'glarea',
 ];
-
-const isEventProp = (name) => /^on[A-Z]/.test(name);
-
-// Props forwarded to ntk createWindow. Event handlers are dispatched by the
-// EventManager from current props (never registered at creation, so they
-// cannot go stale) and children are handled by the tree.
-/**
- * ntk's Window constructor takes every creation attribute up front. The
- * user-facing shape and ntk's differ in two places: size hints are flat
- * props here and a `sizeHints` object there, and the window background is
- * a style property here and a creation attribute there.
- */
-function windowAttributes(props) {
-  const attributes = {};
-  const hints = {};
-  for (const key of Object.keys(props)) {
-    if (key === 'children' || key === 'style' || isEventProp(key)) continue;
-    // resolved in the commit phase against a ref that has not attached yet
-    // at createInstance time — WindowNode._applyTransientFor
-    if (key === 'transientFor') continue;
-    if (WINDOW_HINT_PROPS.includes(key)) {
-      hints[key] = props[key];
-      continue;
-    }
-    attributes[key] = props[key];
-  }
-  if (Object.keys(hints).length > 0) attributes.sizeHints = hints;
-  if (props.style !== undefined) {
-    const style = flattenStyle(props.style);
-    if (style.backgroundColor !== undefined) {
-      attributes.backgroundColor = style.backgroundColor;
-    }
-  }
-  return attributes;
-}
 
 const HostConfig = {
   supportsMutation: true,
